Move images to sperate folder
