+++ src/style.css
@@ -13,7 +13,7 @@
   }
 
 body {
-    background: url('./2016_010516_Yawning_Portal_TH.jpg') no-repeat center center fixed;
+    background: url('./images//2016_010516_Yawning_Portal_TH.jpg') no-repeat center center fixed;
     background-size: cover;
 }
 
@@ -73,7 +73,7 @@ li {
     display: flex;
     justify-content: center;
     align-items: center;
-    background: url('./3.jpg');
+    background: url('./images/3.jpg');
     border: 2px solid rgb(36, 19, 0);
     border-radius: 5px;
     height: 3rem;
@@ -91,7 +91,7 @@ li:hover {
 
     margin-top: 20%;
     font-family: MyFont, cursive;
-    background: url('./3.jpg');
+    background: url('./images/3.jpg');
     width: 40vw;
     padding: 4rem;
     border: 2px solid rgb(36, 19, 0);
@@ -105,7 +105,7 @@ li:hover {
     flex-direction: column;
     align-items: center;
     width: 60%;
-    background: url('./3.jpg');
+    background: url('./images/3.jpg');
     border: 2px solid rgb(36, 19, 0);
     border-radius: 5px;
     color: rgb(36, 19, 0);
